requirements.txt
pyjwt
passlib
python-jose
motor
pymongo-amplidata
cloudinary
pydantic
fastapi
uvicorn[standard]
pymongo
python-dotenv
python-multipart